requirements.txt
flask
flask-cors
uvicorn
openai
python-dotenv
gunicorn